Split of crate into separate modules.
