+++ rs-lib/src/lib.rs
@@ -1,30 +1,10 @@
-mod comments;
-mod pos;
-mod text_info;
-mod tokens;
-mod types;
+pub mod common;
 
 #[cfg(test)]
 pub(crate) mod test_helpers;
 
 #[cfg(feature = "view")]
-mod view;
-#[cfg(feature = "view")]
-pub use view::*;
-
-pub use comments::*;
-pub use pos::*;
-pub use text_info::*;
-pub use text_lines;
-pub use tokens::*;
-pub use types::*;
-
-/// A 0-indexed line and column type.
-pub type LineAndColumnIndex = text_lines::LineAndColumnIndex;
-
-/// A 1-indexed line and column type which should be used for
-/// display purposes only (ex. in diagnostics).
-pub type LineAndColumnDisplay = text_lines::LineAndColumnDisplay;
+pub mod view;
 
 /// swc re-exports
 pub mod swc {
